- Added put method to update task - Changes the refresh token invalid response to bad request with invalid_grant error code - Added Get User Info method to retreive user details - Added key user data to the token response - Added Cors policy to integrate the APIs with Angular site on localhost - Fixed the time related to access token validation

+++ TasksApi/Controllers/TasksController.cs
@@ -53,6 +53,10 @@ namespace TasksApi.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
+            if (id == 0)
+            {
+                return BadRequest(new DeleteTaskResponse { Success = false, ErrorCode = "D01", Error = "Invalid Task id" });
+            }
             var deleteTaskResponse = await taskService.DeleteTask(id, UserID);
             if (!deleteTaskResponse.Success)
             {
@@ -60,6 +64,23 @@ namespace TasksApi.Controllers
             }
 
             return Ok(deleteTaskResponse.TaskId);
+        }
+
+        [HttpPut]
+        public async Task<IActionResult> Put(TaskRequest taskRequest)
+        {
+            var task = new Task { Id = taskRequest.Id, IsCompleted = taskRequest.IsCompleted, Ts = taskRequest.Ts, Name = taskRequest.Name, UserId = UserID };
+
+            var saveTaskResponse = await taskService.SaveTask(task);
+
+            if (!saveTaskResponse.Success)
+            {
+                return UnprocessableEntity(saveTaskResponse);
+            }
+
+            var taskResponse = new TaskResponse { Id = saveTaskResponse.Task.Id, IsCompleted = saveTaskResponse.Task.IsCompleted, Name = saveTaskResponse.Task.Name, Ts = saveTaskResponse.Task.Ts };
+
+            return Ok(taskResponse);
         }
     }
 }
